refactor: both theme files in the custom folder
