+++ src/custom/reactionTheme.js
@@ -1,3 +1,7 @@
+/**
+ * There are two theme files. This is for all MaterialUI components.
+ * The other theme file is `componentTheme.js`
+ */
 import { createMuiTheme } from "@material-ui/core/styles";
 
 const theme = createMuiTheme({
